add indirections in values to allow backpatching semantics for recursive definitions
This approach is very noisy because (un)wrapper for mutable references
must be inserted everywhere in the codebase. On the other hand, it is
also type-directed: it is impossible to forget a pointer dereference
and have a type-correct program. Another approach would be to add
a single `Foward of value option ref` constructor to the type of
values, but then code may silently be incorrect if that case is not
handled properly -- it should be traversed silently by all value
inspections.