lower more: make lower a self-contained pass
The goal of having lowering be a pass of its own is to simplify the
compilation by separating the forms of environments that are required
during lowering (shapes, istail, globals for module fields) and those
that are required during compilation (stack size, locations
for identifiers).

This commit, however, does not perform this separation, it merely
duplicates the environment structures on both sides, in the interest
of simplicity / getting a solution that can be tested relatively
quickly. As a result, the code is arguably worse than before --
sometimes you have to go down a bit before you can go up again.

The simplifications should come in followup commits.