multifile support: guile compile.scm a.ml b.ml c.ml -o foo.byte
The declarations of each file are placed into a module whose name is
derived from the filename in the standard OCaml way.

In particular, this changes slightly the semantics of single-file
invocations: now foo.ml is interpreted as if its declarations were
wrapped into a Foo module. The observable behavior of program does not
change, of course, as we only observe the side-effects of the
evaluation.

(This should not change the compilation either, as structure items,
even nested, are compiled into globals. I checked that the generated
bytecode programs for hello.ml keeps the same size.)